Math: Switch mat_multiply_elementwise product type to int32_t
- Changed product variable from int64_t to int32_t
- Improved performance by reducing data size
- Achieved a 11.57% performance improvement in cycle count

| Rows | Cols | Cycles | Max Error | RMS Error | Result |
+------+------+--------+-----------+-----------+--------+
| 5    | 6    | 2972   | 0.00	   | 0.00      | Pass   |

+++ src/math/matrix.c
@@ -120,7 +120,7 @@ int mat_multiply_elementwise(struct mat_matrix_16b *a, struct mat_matrix_16b *b,
 	int16_t *x = a->data;
 	int16_t *y = b->data;
 	int16_t *z = c->data;
-	int64_t p;
+	int32_t p;
 	int i;
 	const int shift_minus_one = a->fractions + b->fractions - c->fractions - 1;
